+++ react-app/src/index.css
@@ -9,6 +9,7 @@
 
 html {
     font-size: 62.5%;
+
 }
 
 body {
@@ -16,4 +17,10 @@ body {
     font-weight: 400;
     line-height: 1.7;
     color: #262626;
+    background-color: rgb(250, 250, 250);
+
+}
+
+#root {
+    height: 100%;
 }
